+++ src/server/sdb/buffer_writer.cpp
@@ -9,7 +9,7 @@ BufferWriter::BufferWriter(int64_t size)
 
 
 Awaitable<void> BufferWriter::WriteString(const std::string& value) {
-  co_await WriteInt(value.size());
+  co_await WriteInt(static_cast<int64_t>(value.size()));
   ::memcpy(buf_, value.data(), value.size());
   buf_ += value.size();
   co_return;
@@ -21,12 +21,12 @@ Awaitable<void> BufferWriter::WriteInt(int64_t value) {
   co_return;
 }
 
-size_t BufferWriter::EstimateSize(int64_t value) {
+int64_t BufferWriter::EstimateSize(int64_t value) {
   return sizeof(int64_t);
 }
 
-size_t BufferWriter::EstimateSize(const std::string& value) {
-  return sizeof(int64_t) + value.size(); // size + value
+int64_t BufferWriter::EstimateSize(const std::string& value) {
+  return static_cast<int64_t>(sizeof(int64_t)) + static_cast<int64_t>(value.size()); // size + value
 }
 
 std::vector<char> BufferWriter::Extract() && {
