+++ src/server/sdb/writer.cpp
@@ -15,7 +15,7 @@ Awaitable<void> Writer::FSync() {
 }
 
 Awaitable<void> Writer::WriteString(const std::string& value) {
-  co_await WriteInt(value.size());
+  co_await WriteInt(static_cast<int64_t>(value.size()));
   co_await file_writer_->Write(reinterpret_cast<const char*>(value.data()), value.size());
 }
 
@@ -23,12 +23,12 @@ Awaitable<void> Writer::WriteInt(int64_t value) {
   co_await file_writer_->Write(reinterpret_cast<const char*>(&value), sizeof(int64_t));
 }
 
-size_t Writer::EstimateSize(int64_t value) {
+int64_t Writer::EstimateSize(int64_t value) {
   return sizeof(int64_t);
 }
 
-size_t Writer::EstimateSize(const std::string& value) {
-  return sizeof(int64_t) + value.size(); // size + value
+int64_t Writer::EstimateSize(const std::string& value) {
+  return static_cast<int64_t>(sizeof(int64_t)) + static_cast<int64_t>(value.size()); // size + value
 }
 
 }
