+++ src/server/sdb/writer.hpp
@@ -20,10 +20,10 @@ public:
   Awaitable<void> WriteInt(int64_t value);
 
   /// @brief returns size of value after serialization
-  static size_t EstimateSize(int64_t value);
+  static int64_t EstimateSize(int64_t value);
 
   /// @brief returns size of value after serialization
-  static size_t EstimateSize(const std::string& value);
+  static int64_t EstimateSize(const std::string& value);
 private:
   io::FileWriter::Ptr file_writer_;
 };
